Notebook: refactor blueprint business logic to core (#429)
The move of business logic out of the blueprint allows the code to
be reusable without having to go through the web API. It also allows
for tests to be added without the web complexity.

Tests have been refactored so that test clusters will be created only
when testing endpoints that actually require a cluster. This has the
advantage of speeding up test time. This is especially useful when
doing development.

Finally, a new test option has been added so one can use a local
cluster in place of re-creating one each time tests are run.
