+++ .gitignore
@@ -80,4 +80,10 @@ docker-compose.override.yml
 result
 *.qcow2
 
+# tests
 .k3d-config.yaml
+
+# Misc
+*.pem
+*.gz
+*.tgz
