+++ CONTRIBUTING.md
@@ -22,8 +22,7 @@ When reporting a bug, please include the following information:
 If you do not already have a development environment set up, you will probably find the [developer documentation](DEVELOPING.md) helpful.
 
 * Before submitting a pull request, please make sure you agree to the license and have submitted a signed [contributor license agreement](https://github.com/SwissDataScienceCenter/renku/wiki/Legal)
-* PRs should include a short, descriptive title. The titles will be used to compile changelogs for releases, so think about the title in that context. The title should be formatted using the [Conventional Commits](https://www.conventionalcommits.org/) style 
+* PRs should include a short, descriptive title. The titles will be used to compile changelogs for releases, so think about the title in that context. The title should be formatted using the [Conventional Commits](https://www.conventionalcommits.org/) style
 * Small improvements need not reference an issue, but PRs that introduce larger changes or add new functionality should refer to an issue
 * Structure your commits in meaningful units, each with an understandable purpose and coherent commit message. For example, if your proposed changes contain a refactoring and a new feature, make two PRs
 * Format commit messages using the [Conventional Commits](https://www.conventionalcommits.org/) style
-
