+++ bases/renku_data_services/data_api/app.py
@@ -147,7 +147,7 @@ def register_all_handlers(app: Sanic, config: Config) -> Sanic:
         nb_config=config.nb_config,
         project_repo=config.project_repo,
         session_repo=config.session_repo,
-        storage_repo=config.storage_v2_repo,
+        storage_repo=config.storage_repo,
         rp_repo=config.rp_repo,
         internal_gitlab_authenticator=config.gitlab_authenticator,
     )
