+++ components/renku_data_services/notebooks/api/classes/k8s_client.py
@@ -448,8 +448,8 @@ class K8sClient(Generic[_SessionType, _Kr8sType]):
         # NOTE: this get_server ensures the user has access to the server without it you could read someone elses logs
         server = await self.get_server(server_name, safe_username)
         if not server:
-            raise MissingResourceError(
-                f"Cannot find server {server_name} for user " f"{safe_username} to retrieve logs."
+            raise errors.MissingResourceError(
+                message=f"Cannot find server {server_name} for user " f"{safe_username} to retrieve logs."
             )
         pod_name = f"{server_name}-0"
         return await self.renku_ns_client.get_pod_logs(pod_name, max_log_lines)
@@ -490,9 +490,10 @@ class K8sClient(Generic[_SessionType, _Kr8sType]):
         """Delete the server."""
         server = await self.get_server(server_name, safe_username)
         if not server:
-            return None
-        await self.renku_ns_client.delete_server(server_name)
-        return None
+            raise errors.MissingResourceError(
+                message=f"Cannot find server {server_name} for user " f"{safe_username} in order to delete it."
+            )
+        return await self.renku_ns_client.delete_server(server_name)
 
     async def patch_tokens(self, server_name: str, renku_tokens: RenkuTokens, gitlab_token: GitlabToken) -> None:
         """Patch the Renku and Gitlab access tokens used in a session."""
